requirements.txt
huggingface_hub
textblob
nltk
langchain
python-dotenv
chainlit